Bring back Rusoto
In an unintuitive turn of events, AWS Lambda destinations only trigger for asynchronous lambda invocations. SQS does not trigger the function asynchronously, and so an SQS event source mapping cannot be used in conjunction with a destination. More info: https://www.trek10.com/blog/lambda-destinations-what-we-learned-the-hard-way. Bring back Rusoto and call the SendMessage API directly.